+++ assets/components/Calculator.tsx
@@ -1,25 +1,13 @@
 import React, { useState } from 'react';
-import CalculatorContext from "../types/CalculatorContext";
-import DigitButton from './DigitButton';
-import ActionButton from './ActionButton';
-import './../styles/components/Calculator.scss';
-import digits from '../data/digits';
+import Button from './Button';
+import keys from '../data/keys';
 import computation from '../data/computation';
+import './../styles/components/Calculator.scss';
 
 const Calculator: React.FC = () => {
     const [calculatorInput, setCalculatorInput] = useState<string>('');
     const [stateQueue, setStateQueue] = useState<string[]>([]);
     const [clearNext, setClearNext] = useState<boolean>(false);
-
-    const revertToPreviousState = (): void => setCalculatorInput(stateQueue.pop() || '');
-
-    const compute = async (): Promise<void> => {
-        QueueInput();
-        
-        const result: string = await computation(calculatorInput);
-
-        setCalculatorInput(result);
-    };
 
     const enterInput = (newInput: string): void => {
         clearNext && QueueInput();
@@ -32,29 +20,33 @@ const Calculator: React.FC = () => {
 
     const QueueInput = (state: string = calculatorInput): void => setStateQueue([...stateQueue, state]);
 
-    const clearInput = (): void => {
-        clearNext && QueueInput();
+    const clean = (): void => setCalculatorInput(stateQueue.pop() || '');
+
+    const reset = (): void => {
+        setStateQueue([]);
         setCalculatorInput('');
+    };
+
+    const compute = async (): Promise<void> => {
+        QueueInput();
+        
+        const result: string = await computation(calculatorInput);
+
+        setCalculatorInput(result);
     };
     
     return (
-        <CalculatorContext.Provider value={{ enterInput }}>
-            <section className='calculator'>
-                <div className='calculator__display'>
-                    <div className="display__history"><span className='display__text'>{ stateQueue }</span></div>
-                    <div className="display__current">
-                        <span className='display__text'>{ calculatorInput }</span>
-                    </div>
+        <section className='calculator'>
+            <div className='calculator__display'>
+                <div className="display__history"><span className='display__text'>{ stateQueue }</span></div>
+                <div className="display__current">
+                    <span className='display__text'>{ calculatorInput }</span>
                 </div>
-                <div className='calculator__pad'>
-                    { digits.map(digit => <DigitButton key={ digit.symbol } digit={ digit } perform={ enterInput } />) }
-                    <ActionButton symbol='C' slug='clear' action={ revertToPreviousState } />
-                    <ActionButton symbol='AC' slug='reset' action={ clearInput } />
-                    <ActionButton symbol='=' slug='equal' action={ compute } />
-                    {/* { Actions.map(Action => { return <ActionButton action={Action} perform={compute} /> })}  */}
-                </div>
-            </section>
-        </CalculatorContext.Provider>
+            </div>
+            <div className='calculator__pad'>
+                { keys.map(key => <Button key={ key.symbol } digit={ key } perform={ key.perform ? eval(key.perform) : enterInput } />) }
+            </div>
+        </section>
     );
 }
