+++ assets/styles/components/Calculator.scss
@@ -8,7 +8,7 @@
     margin-bottom: 10px;
     border: 1px solid #ebebeb;
     border-radius: 8px;
-
+    
     .display__history {
       margin: 8px 0 5px 0;
       height: 20px;
