--- assets/types/Digit.ts
@@ -1,7 +1,0 @@
-type Digit = {
-    symbol: string,
-    slug: string,
-    keyCode: string
-}
-
-export default Digit
